Refactor

+++ tests/TestHelper.h
@@ -5,11 +5,10 @@
 #include <type_traits>
 
 template <typename... Args>
-std::tuple<size_t, char**> createArgcArgv(Args... args) {
+std::tuple<int, char**> createArgcArgv(Args... args) {
   const size_t N = sizeof...(Args);
   char** array = new char*[N];
   size_t i = 0;
   (..., (array[i++] = strdup(args)));
   return std::make_tuple(N, array);
 }
-
